Added new chart type which is based on column chart.
Column height and width are both read from data. Width measure is
specified in summary presentation and it is added as a 'normal' measure
to summary so it appears correctly in data url. Svg drawing is based on
column chart but was separated from it to keep code cleaner. We have to
calculate column widths and locations before actual drawing, because of
the difficulties with matching column sizes with x-axis tick positions
(especially when data is sorted based on column heights).

Issue PIVOT-660